fix(integration-test): separate dswx-s1 tests
Refs #705
