+++ pip.conf
@@ -9,3 +9,4 @@
 [global]
 # passthrough proxy and cache for PyPI.
 index-url = https://artifactory.jpl.nasa.gov/artifactory/api/pypi/pypi-release-virtual/simple
+extra-index-url = https://pypi.org/simple
